requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
attrs==19.3.0
Automat==0.8.0
billiard==3.6.4.0
blinker==1.4
cached-property==1.5.2
celery==5.1.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloud-init==20.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.5
django-baton==2.2.1
entrypoints==0.3
filelock==3.0.12
httplib2==0.19.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.6.1
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.1.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
oauth2client==1.5.2
oauthlib==3.1.0
pexpect==4.6.0
prompt-toolkit==3.0.19
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-debian===0.1.36ubuntu1
pytz==2021.1
PyYAML==5.3.1
redis==3.5.3
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.7.2
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sos==4.0
sqlparse==0.4.1
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
typing-extensions==3.10.0.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vine==5.0.0
virtualenv==20.4.3
wadllib==1.3.3
wcwidth==0.2.5
zipp==3.5.0
zope.interface==4.7.1
